# docs/swagger.yaml
basePath: /
definitions:
  handler.CreateUpdateSubscriptionRequest:
    properties:
      end_date:
        example: 10-12-2027
        type: string
      price:
        example: 400
        type: integer
      service_name:
        example: Yandex Plus
        type: string
      start_date:
        example: 07-01-2025
        type: string
      user_id:
        example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
        type: string
    type: object
  handler.SubscriptionCreateResponse:
    properties:
      message:
        example: подписка успешно создана
        type: string
      subscription:
        $ref: '#/definitions/model.SubscriptionDetails'
    type: object
  handler.SubscriptionUpdateResponse:
    properties:
      message:
        example: подписка успешно обновлена
        type: string
    type: object
  handler.TotalCostResponse:
    properties:
      user_id:
        type: string
      общая_стоимость:
        type: integer
    type: object
  model.SubscriptionDetails:
    properties:
      end_date:
        type: string
      id:
        type: integer
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Сервис управления подписками
  title: Subscription API
  version: "1.0"
paths:
  /subscriptions/create:
    post:
      consumes:
      - application/json
      description: Создаёт новую подписку с указанием пользователя, сервиса, стоимости
        и периода
      parameters:
      - description: Детали подписки
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/handler.CreateUpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.SubscriptionCreateResponse'
        "400":
          description: неверный формат запроса
          schema:
            type: string
        "500":
          description: ошибка создания подписки
          schema:
            type: string
      summary: Создание подписки
      tags:
      - Подписки
  /subscriptions/delete/{id}:
    delete:
      description: Удаляет подписку по её идентификатору
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: подписка успешно удалена
        "400":
          description: неверный ID
          schema:
            type: string
        "500":
          description: не удалось удалить подписку
          schema:
            type: string
      summary: Удалить подписку по ID
      tags:
      - Подписки
  /subscriptions/get/{id}:
    get:
      description: Возвращает подписку по её уникальному идентификатору
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SubscriptionDetails'
        "400":
          description: неверный ID
          schema:
            type: string
        "404":
          description: не удалось поулчить подписку
          schema:
            type: string
      summary: Получение подписки по ID
      tags:
      - Подписки
  /subscriptions/total-cost:
    get:
      description: Возвращает сумму всех подписок пользователя с возможностью фильтрации
        по сервису и диапазону дат
      parameters:
      - description: UUID пользователя
        in: query
        name: user_id
        required: true
        type: string
      - description: Название сервиса (опционально)
        in: query
        name: service_name
        type: string
      - description: Дата начала (формат DD-MM-YYYY, по умолчанию 01-01-2000)
        in: query
        name: start_date
        type: string
      - description: Дата окончания (формат DD-MM-YYYY)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.TotalCostResponse'
        "400":
          description: ошибка параметров запроса
          schema:
            type: string
        "500":
          description: ошибка сервера
          schema:
            type: string
      summary: Получение общей стоимости подписок пользователя
      tags:
      - Подписки
  /subscriptions/update/{id}:
    put:
      consumes:
      - application/json
      description: Обновляет информацию о подписке по заданному ID
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      - description: Обновлённые данные подписки
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/handler.CreateUpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.SubscriptionUpdateResponse'
        "400":
          description: неверный ID или формат запроса
          schema:
            type: string
        "500":
          description: не удалось обновить информацию по подписке
          schema:
            type: string
      summary: Обновить подписку по ID
      tags:
      - Подписки
  /subscriptions/user/{uuid}:
    get:
      description: Возвращает список всех подписок по UUID пользователя
      parameters:
      - description: UUID пользователя
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.SubscriptionDetails'
            type: array
        "404":
          description: не удалось получить подписки
          schema:
            type: string
      summary: Получение подписок пользователя
      tags:
      - Подписки
schemes:
- http
swagger: "2.0"